+++ backend/canisters/group/api/src/updates/c2c_join_group.rs
@@ -8,6 +8,8 @@ pub struct Args {
     pub principal: Principal,
     pub invite_code: Option<u64>,
     pub correlation_id: u64,
+    #[serde(default)]
+    pub is_platform_moderator: bool,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]