+++ backend/canisters/group/impl/src/queries/public_summary.rs
@@ -31,7 +31,6 @@ fn public_summary_impl(args: Args, runtime_state: &RuntimeState) -> Response {
         latest_message: events_reader.latest_message_event(None),
         latest_event_index,
         participant_count: data.participants.len(),
-        owner_id: runtime_state.data.owner_id,
         is_public: runtime_state.data.is_public,
         frozen: runtime_state.data.frozen.value.clone(),
         events_ttl: runtime_state.data.events.get_events_time_to_live().value,