+++ backend/canisters/group_index/impl/src/model/public_groups.rs
@@ -100,7 +100,6 @@ impl PublicGroups {
             latest_message: summary.latest_message,
             latest_event_index: summary.latest_event_index,
             participant_count: summary.participant_count,
-            owner_id: summary.owner_id,
             is_public: true,
             frozen: None,
             events_ttl: summary.events_ttl,