+++ backend/canisters/proposals_bot/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Fixed
+
+- Call group::change_role using candid ([#3340](https://github.com/open-ic/open-chat/pull/3340))
+
 ## [[2.0.638](https://github.com/open-ic/open-chat/releases/tag/v2.0.638-proposals_bot)] - 2023-03-20
 
 ### Added